+++ the_keep/views.py
@@ -1212,9 +1212,10 @@ def ultimate_component_view(request, slug, component):
             color_label = None
             object_color = None
 
-    absolute_uri = build_absolute_uri(request, obj.get_absolute_url())
+    absolute_uri = f'{build_absolute_uri(request, obj.get_absolute_url())}?lang={language_code}'
 
     # Meta Tags
+    meta_title = object_title
     meta_description = obj.get_meta_description(language)
     meta_image_url = obj.get_meta_image_url(language)
 
@@ -1291,6 +1292,7 @@ def ultimate_component_view(request, slug, component):
         'wr_link_config': wr_link_config,
         'fr_link_config': fr_link_config,
 
+        'meta_title': meta_title,
         'meta_description': meta_description,
         'meta_image_url': meta_image_url,
 
@@ -1311,8 +1313,8 @@ def ultimate_component_view(request, slug, component):
 
 def color_group_view(request, color_name):
 
-    language = get_language()
-    language_object = Language.objects.filter(code=language).first()
+    language_code = get_language()
+    language_object = Language.objects.filter(code=language_code).first()
 
     color_name = str(color_name).capitalize()
     color_group = ColorChoices.get_color_by_name(color_name)
@@ -1372,8 +1374,8 @@ def color_group_view(request, color_name):
 
 def animal_match_view(request, slug):
 
-    language = get_language()
-    language_object = Language.objects.filter(code=language).first()
+    language_code = get_language()
+    language_object = Language.objects.filter(code=language_code).first()
     # post = get_object_or_404(Post, slug=slug)
     post = Post.objects.filter(slug=slug).annotate(
                 selected_title=Coalesce(
@@ -1439,8 +1441,6 @@ def animal_match_view(request, slug):
                     'lore'  # Fall back to the default lore if there's no translation
                 )
             ).distinct()
-    language = get_language()
-    language_object = Language.objects.filter(code=language).first()
     # Try to get the translated animal name from PostTranslation
     translated_animal = None
     if language_object:
@@ -1533,8 +1533,14 @@ def component_games(request, slug, component):
     except EmptyPage:
         page_obj = paginator.page(paginator.num_pages)  # Redirect to the last page if invalid
 
+    language_code = get_language()
+    language = Language.objects.filter(code=language_code).first()
 
-
+    # Meta Tags
+    absolute_uri = build_absolute_uri(request, obj.get_absolute_url())
+    meta_title = f'{obj.title} Games'
+    meta_description = f'Games recorded with {obj.title}'
+    meta_image_url = obj.get_meta_image_url(language)
 
     context = {
         'object': obj,
@@ -1549,6 +1555,11 @@ def component_games(request, slug, component):
         'win_rate': win_rate,
         'tourney_points': tourney_points,
         'total_efforts': total_efforts,
+
+        'absolute_uri': absolute_uri,
+        'meta_title': meta_title,
+        'meta_description': meta_description,
+        'meta_image_url': meta_image_url,
     }
     if request.htmx:
             return render(request, 'the_keep/partials/game_list.html', context)
@@ -1559,8 +1570,8 @@ def component_games(request, slug, component):
 
 @login_required
 @bookmark_toggle(Post)
-def bookmark_post(request, object):
-    return render(request, 'the_keep/partials/bookmarks.html', {'object': object })
+def bookmark_post(request, obj):
+    return render(request, 'the_keep/partials/bookmarks.html', {'object': obj})
 
 
 
@@ -2805,9 +2816,9 @@ def status_check(request, slug):
     Klass = COMPONENT_MAPPING.get(post.component)
     obj = get_object_or_404(Klass, slug=slug)
 
-    language = get_language()
-    language_object = Language.objects.filter(code=language).first()
-    object_translation = obj.translations.filter(language=language_object).first()
+    language_code = get_language()
+    language = Language.objects.filter(code=language_code).first()
+    object_translation = obj.translations.filter(language=language).first()
     object_title = object_translation.translated_title if object_translation and object_translation.translated_title else obj.title
 
 
@@ -2912,6 +2923,12 @@ def status_check(request, slug):
                 object_color = "#5f788a"
         else:
             object_color = "#5f788a"
+    
+    # Meta Tags
+    absolute_uri = build_absolute_uri(request, obj.get_absolute_url())
+    meta_title = f'{obj.title} Status'
+    meta_description = f'{total_completion} Progress towards Stable Status ({total_count}/{total_threshold})'
+    meta_image_url = obj.get_meta_image_url(language)
 
     context = {
         'object': obj,
@@ -2958,6 +2975,11 @@ def status_check(request, slug):
         'total_count': total_count,
         'total_threshold': total_threshold,
         'total_completion': total_completion,
+
+        'absolute_uri': absolute_uri,
+        'meta_title': meta_title,
+        'meta_description': meta_description,
+        'meta_image_url': meta_image_url,
     }
 
     return render(request, 'the_keep/status_check.html', context)
@@ -3174,6 +3196,7 @@ class PNPAssetListView(ListView):
     def get_context_data(self, **kwargs):
         # Add current user to the context data
         context = super().get_context_data(**kwargs)
+        context['active_page'] = 'workshop'  # Set active page for navigation
         if self.request.user.is_authenticated:
             context['profile'] = self.request.user.profile  # Adding the user to the context
             context['shared_assets'] = PNPAsset.objects.filter(shared_by__slug=self.request.user.profile.slug)
@@ -3236,7 +3259,58 @@ class PNPAssetListView(ListView):
         #     send_discord_message(f'{get_uuid(self.request)} viewing The Workshop')
 
         return super().render_to_response(context, **response_kwargs)
-    
+
+
+class MyPNPAssetListView(LoginRequiredMixin, ListView):
+    model = PNPAsset
+    template_name = 'the_keep/asset_list.html'
+    context_object_name = 'objects'
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        # Filter for assets shared by the current logged-in user
+        queryset = queryset.filter(shared_by=self.request.user.profile)
+
+        # Get the search query from the GET parameters
+        search_query = self.request.GET.get('search', '')
+        search_type = self.request.GET.get('search_type', '')
+        file_type = self.request.GET.get('file_type', '')
+
+        # If a search query is provided, filter the queryset
+        if search_query:
+            queryset = queryset.filter(Q(title__icontains=search_query))
+        if search_type:
+            queryset = queryset.filter(category__icontains=search_type)
+        if file_type:
+            queryset = queryset.filter(file_type__icontains=file_type)
+
+        return queryset
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context['active_page'] = 'my-resources'  # Set active page for navigation
+        context['profile'] = self.request.user.profile
+        context['shared_assets'] = self.get_queryset()
+
+        theme = get_theme(self.request)
+        background_image, foreground_images, theme_artists, background_pattern = get_thematic_images(theme=theme, page='resources')
+        context['background_image'] = background_image
+        context['foreground_images'] = foreground_images
+        context['background_pattern'] = background_pattern
+
+        # Add unpinned_assets as empty for consistency with the template
+        context['unpinned_assets'] = PNPAsset.objects.none()
+
+        return context
+
+    def render_to_response(self, context, **response_kwargs):
+        # Check if it's an HTMX request
+        if self.request.headers.get('HX-Request') == 'true':
+            return render(self.request, 'the_keep/partials/asset_list_table.html', context)
+
+        return super().render_to_response(context, **response_kwargs)
+
+
 @player_required_class_based_view
 class PNPAssetDeleteView(LoginRequiredMixin, UserPassesTestMixin, DeleteView):
     model = PNPAsset
@@ -3280,7 +3354,7 @@ def pin_asset(request, id):
         obj.pinned = True
     obj.save()
 
-    return render(request, 'the_keep/partials/asset_pins.html', {'obj': object })
+    return render(request, 'the_keep/partials/asset_pins.html', {'obj': obj})
 
 
 
@@ -3319,8 +3393,8 @@ def universal_search(request):
     translations = PostTranslation.objects.none()
     translated_posts = Post.objects.none()
 
-    language = get_language()
-    language_object = Language.objects.filter(code=language).first()
+    language_code = get_language()
+    language_object = Language.objects.filter(code=language_code).first()
 
     if not query:
         factions = Faction.objects.none()
@@ -3341,6 +3415,7 @@ def universal_search(request):
         laws = Law.objects.none()
         bot_laws = Law.objects.none()
         fan_laws = Law.objects.none()
+        cards = Card.objects.none()
     else:
         # If the query is not empty, perform the search as usual
         factions = search_translatable_posts(Faction, query, language_object, view_status)
@@ -3375,7 +3450,7 @@ def universal_search(request):
             
         # Laws, look for exact matches first, then contains matches
         laws = Law.objects.filter(
-            language__code=language,
+            language__code=language_code,
             group__type__in=['Official', 'Appendix'],
             group__public=True
         ).filter(
@@ -3383,7 +3458,7 @@ def universal_search(request):
         )
         if not laws:
             laws = Law.objects.filter(
-                language__code=language,
+                language__code=language_code,
                 group__type__in=['Official', 'Appendix'],
                 group__public=True
             ).filter(
@@ -3391,7 +3466,7 @@ def universal_search(request):
             )
 
         bot_laws = Law.objects.filter(
-            language__code=language,
+            language__code=language_code,
             group__type='Bot',
             group__public=True
         ).filter(
@@ -3399,7 +3474,7 @@ def universal_search(request):
         )
         if not bot_laws:
             bot_laws = Law.objects.filter(
-                language__code=language,
+                language__code=language_code,
                 group__type='Bot',
                 group__public=True
             ).filter(
@@ -3407,7 +3482,7 @@ def universal_search(request):
             )
 
         fan_laws = Law.objects.filter(
-            language__code=language,
+            language__code=language_code,
             group__type='Fan',
             group__public=True
         ).filter(
@@ -3415,12 +3490,13 @@ def universal_search(request):
         )
         if not fan_laws:
             fan_laws = Law.objects.filter(
-                language__code=language,
+                language__code=language_code,
                 group__type='Fan',
                 group__public=True
             ).filter(
                 (Q(plain_title__icontains=query))
             )
+        cards = Card.objects.filter(name__icontains=query)
 
     if color_group:
         color_count = 1
@@ -3432,7 +3508,8 @@ def universal_search(request):
                      players.count() + games.count() + scorecards.count() + 
                      tournaments.count() + rounds.count() + tweaks.count() + 
                      resources.count() + pieces.count() + color_count + translations.count() + translated_posts.count() +
-                     laws.count() + bot_laws.count() + fan_laws.count()
+                     laws.count() + bot_laws.count() + fan_laws.count() +
+                     cards.count()
                      )
     
     no_results = total_results == 0
@@ -3466,6 +3543,7 @@ def universal_search(request):
         'laws': laws[:result_count],
         'bot_laws': bot_laws[:result_count],
         'fan_laws': fan_laws[:result_count],
+        'cards': cards[:result_count],
         'no_results': no_results,
         'query': query,
     }
@@ -5097,14 +5175,34 @@ def view_deckgroup(request, post_slug, language_code, deckgroup_slug):
     language = get_object_or_404(Language, code=language_code)
     has_only_one_card = post.has_only_one_card()
     deckgroup = get_object_or_404(
-        DeckGroup, 
-        slug=deckgroup_slug, 
-        post=post, 
+        DeckGroup,
+        slug=deckgroup_slug,
+        post=post,
         language=language
     )
- 
+
     can_edit = user_can_edit(request, post)
     title = deckgroup.name
+    highlight_card_id = request.GET.get('highlight_card')
+    highlighted_card = None
+    absolute_uri = None
+    meta_title = f'{deckgroup.name} - {deckgroup.post.title}'
+    meta_description = f'{deckgroup.card_count} {deckgroup.name} for {deckgroup.post.title}'
+    meta_image_url = deckgroup.back_image.url
+    if highlight_card_id:
+        highlighted_card = Card.objects.filter(group=deckgroup, id=highlight_card_id).first()
+        if highlighted_card:
+            print('highlight')
+            absolute_uri = build_absolute_uri(request, highlighted_card.get_absolute_url())
+            meta_image_url = highlighted_card.front_image.url
+            if highlighted_card.name:
+                meta_title = f'{highlighted_card.name} - {deckgroup.name}'
+            if highlighted_card.text:
+                meta_description = highlighted_card.text
+
+    if not absolute_uri:
+        absolute_uri = build_absolute_uri(request, deckgroup.get_absolute_url())
+
     context = {
         "deckgroup": deckgroup,
         "post": post,
@@ -5113,6 +5211,12 @@ def view_deckgroup(request, post_slug, language_code, deckgroup_slug):
         'language_code': language_code,
         "title": title,
         "has_only_one_card": has_only_one_card,
+        "highlight_card_id": highlight_card_id,
+
+        "absolute_uri": absolute_uri,
+        "meta_title": meta_title,
+        "meta_description": meta_description,
+        "meta_image_url": meta_image_url,
         }
     return render(request, "the_keep/deckgroup/detail.html", context)